pyproject.toml
[project]
name = "chat-ontology-builder-backend"
version = "0.1.0"
description = "Backend for Kantian Self Learning Ontologies"
authors = [
    {name = "Allan Lund Hansen",email = "allanlundhansen@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "neo4j (>=5.28.1,<6.0.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.0"
pytest = "^8.3.5"
httpx = "^0.28.1"
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
python_files = "tests.py test_*.py *_tests.py"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
log_cli = true
log_cli_level = "INFO"
markers = [
    "anyio",
]
filterwarnings = [
    # Example: Ignore specific warnings if needed
    # "ignore::DeprecationWarning",
]
